refactor: :recycle: 파일 경로 수정 및 절대경로 refactor
