+++ src/pages/Consult/components/recording/AiSummary.tsx
@@ -3,7 +3,7 @@ import React from 'react';
 const AiSummary: React.FC = () => {
   return (
     <div className="mt-2 h-[720px] border-[1px] border-grayscale-30 rounded-[4px] p-1">
-      asdf
+      AI요약
     </div>
   );
 };
